+++ lib/src/popover.dart
@@ -21,7 +21,7 @@ class Popover extends Tooltip {
   Popover(Element element, {bool animation, String placement(Element elem), 
   String selector, String template, String trigger, String title(Element elem), 
   String content(Element elem), int delay, int showDelay, 
-  int hideDelay, bool html, container}) : 
+  int hideDelay, bool html, container, NodeValidatorBuilder htmlValidator}) : 
   this.template = _data(template, element, 'template', _DEFAULT_TEMPLATE),
   this.trigger  = _data(trigger,  element, 'trigger',  'click'),
   this._content = p.fallback(content, () => (Element elem) => elem.attributes['data-content']),