+++ lib/src/tooltip.dart
@@ -185,7 +185,7 @@ class Tooltip extends Base {
     if (container != null)
       $(tip).appendTo(container);
     else
-      $element.after('body');
+      $element.after(tip);
     
     final Rectangle pos = _position;
     final int actualWidth = tip.offsetWidth;